[diffusion] merge wan s2v dit implementation
